device/requirements.txt
psutil>=5.9.0
paho-mqtt>=2.0.0

